Enable Flutter Gold in Fuchsia driver test
