+++ dev/devicelab/lib/tasks/integration_tests.dart
@@ -107,15 +107,22 @@ TaskFunction createAndroidSplashScreenKitchenSinkTest() {
   );
 }
 
-TaskFunction createFlutterDriverScreenshotTest() {
+/// Executes a driver test that takes a screenshot and compares it against a golden image.
+/// If [useFlutterGold] is true, the golden image is served by Flutter Gold
+/// (https://flutter-gold.skia.org/), otherwise the golden image is read from the disk.
+TaskFunction createFlutterDriverScreenshotTest({
+  bool useFlutterGold = false,
+}) {
   return DriverTest(
     '${flutterDirectory.path}/dev/integration_tests/flutter_driver_screenshot_test',
     'lib/main.dart',
+    extraOptions: useFlutterGold ? const <String>[
+      '--driver', 'test_driver/flutter_gold_main_test.dart'
+    ] : null
   );
 }
 
 class DriverTest {
-
   DriverTest(
     this.testDirectory,
     this.testTarget, {
