+++ dev/integration_tests/flutter_driver_screenshot_test/README.md
@@ -4,7 +4,12 @@ This tests contains an app with a main page and sub pages.
 The main page contains a list of buttons; each button leads to a designated sub page when tapped on.
 Each sub page should displays some simple UIs to screenshot tested.
 
-The flutter driver test runs the app and opens each page to take a screenshot. Then it compares the screenshot against a golden image stored in `test_driver/goldens/<some_test_page_name>/<device_model>.png`.
+The flutter driver test runs the app and opens each page to take a screenshot.
+
+Use `test_driver/flutter_gold_main_test.dart` to test against golden files stored on Flutter Gold.
+Otherwise, use `main_test.dart` to test against golden files stored on `test_driver/goldens/<some_test_page_name>/<device_model>.png`.
+
+Note that new binaries can't be checked in the Flutter repo, so use [Flutter Gold](https://github.com/flutter/flutter/wiki/Writing-a-golden-file-test-for-package:flutter) instead.
 
 # Add a new page to test
 
@@ -16,6 +21,7 @@ The flutter driver test runs the app and opens each page to take a screenshot. T
 
 An example of a `Page` subclass can be found in `lib/image_page.dart`
 
-# Experiments
+# Environments
 
-The test currently only runs on device lab ["mac/ios"] which runs the app on iPhone 6s.
+* Device Lab which runs the app on iPhone 6s.
+* LUCI which runs the app on a Fuchsia NUC device.
