+++ packages/flutter_test/lib/src/goldens.dart
@@ -3,12 +3,12 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:math' as math;
 import 'dart:typed_data';
 
-import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart';
+import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path;
-import '_goldens_io.dart' if (dart.library.html) '_goldens_web.dart' as _goldens;
+import 'package:image/image.dart';
 
 /// Compares image pixels against a golden image file.
 ///
@@ -98,7 +98,77 @@ abstract class GoldenFileComparator {
   /// Returns a [ComparisonResult] to describe the pixel differential of the
   /// [test] and [master] image bytes provided.
   static ComparisonResult compareLists(List<int> test, List<int> master) {
-    return _goldens.compareLists(test, master);
+    if (identical(test, master))
+      return ComparisonResult(passed: true);
+
+    if (test == null || master == null || test.isEmpty || master.isEmpty) {
+      return ComparisonResult(
+        passed: false,
+        error: 'Pixel test failed, null image provided.',
+      );
+    }
+
+    final Image testImage = decodePng(test);
+    final Image masterImage = decodePng(master);
+
+    assert(testImage != null);
+    assert(masterImage != null);
+
+    final int width = testImage.width;
+    final int height = testImage.height;
+
+    if (width != masterImage.width || height != masterImage.height) {
+      return ComparisonResult(
+        passed: false,
+        error: 'Pixel test failed, image sizes do not match.\n'
+          'Master Image: ${masterImage.width} X ${masterImage.height}\n'
+          'Test Image: ${testImage.width} X ${testImage.height}',
+      );
+    }
+
+    int pixelDiffCount = 0;
+    final int totalPixels = width * height;
+    final Image invertedMaster = invert(Image.from(masterImage));
+    final Image invertedTest = invert(Image.from(testImage));
+
+    final Map<String, Image> diffs = <String, Image>{
+      'masterImage' : masterImage,
+      'testImage' : testImage,
+      'maskedDiff' : Image.from(testImage),
+      'isolatedDiff' : Image(width, height),
+    };
+
+    for (int x = 0; x < width; x++) {
+      for (int y = 0; y < height; y++) {
+        final int testPixel = testImage.getPixel(x, y);
+        final int masterPixel = masterImage.getPixel(x, y);
+
+        final int diffPixel = (getRed(testPixel) - getRed(masterPixel)).abs()
+          + (getGreen(testPixel) - getGreen(masterPixel)).abs()
+          + (getBlue(testPixel) - getBlue(masterPixel)).abs()
+          + (getAlpha(testPixel) - getAlpha(masterPixel)).abs();
+
+        if (diffPixel != 0 ) {
+          final int invertedMasterPixel = invertedMaster.getPixel(x, y);
+          final int invertedTestPixel = invertedTest.getPixel(x, y);
+          final int maskPixel = math.max(invertedMasterPixel, invertedTestPixel);
+          diffs['maskedDiff'].setPixel(x, y, maskPixel);
+          diffs['isolatedDiff'].setPixel(x, y, maskPixel);
+          pixelDiffCount++;
+        }
+      }
+    }
+
+    if (pixelDiffCount > 0) {
+      return ComparisonResult(
+        passed: false,
+        error: 'Pixel test failed, '
+          '${((pixelDiffCount/totalPixels) * 100).toStringAsFixed(2)}% '
+          'diff detected.',
+        diffs: diffs,
+      );
+    }
+    return ComparisonResult(passed: true);
   }
 }
 
@@ -133,115 +203,6 @@ GoldenFileComparator _goldenFileComparator = const TrivialComparator._();
 set goldenFileComparator(GoldenFileComparator value) {
   assert(value != null);
   _goldenFileComparator = value;
-}
-
-/// Compares image pixels against a golden image file.
-///
-/// Instances of this comparator will be used as the backend for
-/// [matchesGoldenFile] when tests are running on Flutter Web, and will usually
-/// implemented by deferring the screenshot taking and image comparison to a
-/// test server.
-///
-/// Instances of this comparator will be invoked by the test framework in the
-/// [TestWidgetsFlutterBinding.runAsync] zone and are thus not subject to the
-/// fake async constraints that are normally imposed on widget tests (i.e. the
-/// need or the ability to call [WidgetTester.pump] to advance the microtask
-/// queue). Prior to the invocation, the test framework will render only the
-/// [Element] to be compared on the screen.
-///
-/// See also:
-///
-///  * [GoldenFileComparator] for the comparator to be used when the test is
-///    not running in a web browser.
-///  * [DefaultWebGoldenComparator] for the default [WebGoldenComparator]
-///    implementation for `flutter test`.
-///  * [matchesGoldenFile], the function from [flutter_test] that invokes the
-///    comparator.
-abstract class WebGoldenComparator {
-  /// Compares the rendered pixels of [element] of size [size] that is being
-  /// rendered on the top left of the screen against the golden file identified
-  /// by [golden].
-  ///
-  /// The returned future completes with a boolean value that indicates whether
-  /// the pixels rendered on screen match the golden file's pixels.
-  ///
-  /// In the case of comparison mismatch, the comparator may choose to throw a
-  /// [TestFailure] if it wants to control the failure message, often in the
-  /// form of a [ComparisonResult] that provides detailed information about the
-  /// mismatch.
-  ///
-  /// The method by which [golden] is located and by which its bytes are loaded
-  /// is left up to the implementation class. For instance, some implementations
-  /// may load files from the local file system, whereas others may load files
-  /// over the network or from a remote repository.
-  Future<bool> compare(Element element, Size size, Uri golden);
-
-  /// Updates the golden file identified by [golden] with rendered pixels of
-  /// [element].
-  ///
-  /// This will be invoked in lieu of [compare] when [autoUpdateGoldenFiles]
-  /// is `true` (which gets set automatically by the test framework when the
-  /// user runs `flutter test --update-goldens --platform=chrome`).
-  ///
-  /// The method by which [golden] is located and by which its bytes are written
-  /// is left up to the implementation class.
-  Future<void> update(Uri golden, Element element, Size size);
-
-  /// Returns a new golden file [Uri] to incorporate any [version] number with
-  /// the [key].
-  ///
-  /// The [version] is an optional int that can be used to differentiate
-  /// historical golden files.
-  ///
-  /// Version numbers are used in golden file tests for package:flutter. You can
-  /// learn more about these tests [here](https://github.com/flutter/flutter/wiki/Writing-a-golden-file-test-for-package:flutter).
-  Uri getTestUri(Uri key, int version) {
-    if (version == null)
-      return key;
-    final String keyString = key.toString();
-    final String extension = path.extension(keyString);
-    return Uri.parse(
-      keyString
-        .split(extension)
-        .join() + '.' + version.toString() + extension
-    );
-  }
-}
-
-/// Compares pixels against those of a golden image file.
-///
-/// This comparator is used as the backend for [matchesGoldenFile] when tests
-/// are running in a web browser.
-///
-/// When using `flutter test --platform=chrome`, a comparator implemented by
-/// [DefaultWebGoldenComparator] is used if no other comparator is specified. It
-/// will send a request to the test server, which uses [goldenFileComparator]
-/// for golden file compatison.
-///
-/// When using `flutter test --update-goldens`, the [DefaultWebGoldenComparator]
-/// updates the files on disk to match the rendering.
-///
-/// When using `flutter run`, the default comparator
-/// ([_TrivialWebGoldenComparator]) is used. It prints a message to the console
-/// but otherwise does nothing. This allows tests to be developed visually on a
-/// web browser.
-///
-/// Callers may choose to override the default comparator by setting this to a
-/// custom comparator during test set-up (or using directory-level test
-/// configuration). For example, some projects may wish to install a comparator
-/// with tolerance levels for allowable differences.
-///
-/// See also:
-///
-///  * [flutter_test] for more information about how to configure tests at the
-///    directory-level.
-///  * [goldenFileComparator], the comparator used when tests are not running on
-///    a web browser.
-WebGoldenComparator get webGoldenComparator => _webGoldenComparator;
-WebGoldenComparator _webGoldenComparator = const _TrivialWebGoldenComparator._();
-set webGoldenComparator(WebGoldenComparator value) {
-  assert(value != null);
-  _webGoldenComparator = value;
 }
 
 /// Whether golden files should be automatically updated during tests rather
@@ -280,33 +241,13 @@ class TrivialComparator implements GoldenFileComparator {
 
   @override
   Future<bool> compare(Uint8List imageBytes, Uri golden) {
-    debugPrint('Golden file comparison requested for "$golden"; skipping...');
+    print('Golden file comparison requested for "$golden"; skipping...');
     return Future<bool>.value(true);
   }
 
   @override
   Future<void> update(Uri golden, Uint8List imageBytes) {
     throw StateError('goldenFileComparator has not been initialized');
-  }
-
-  @override
-  Uri getTestUri(Uri key, int version) {
-    return key;
-  }
-}
-
-class _TrivialWebGoldenComparator implements WebGoldenComparator {
-  const _TrivialWebGoldenComparator._();
-
-  @override
-  Future<bool> compare(Element element, Size size, Uri golden) {
-    debugPrint('Golden comparison requested for "$golden"; skipping...');
-    return Future<bool>.value(true);
-  }
-
-  @override
-  Future<void> update(Uri golden, Element element, Size size) {
-    throw StateError('webGoldenComparator has not been initialized');
   }
 
   @override
